move all common scheduelr functionality to SharedScheduler
